+++ app/src/main/java/com/yifarj/yifadinghuobao/database/model/CollectionItemModel.java
@@ -46,7 +46,7 @@ public class CollectionItemModel extends BaseRXModel {
     public String BatchId;
 
     @Column
-    public int Quantity;
+    public double Quantity;
 
     @Column
     public double BasicQuantity;
